Refactors encoded (binary) record headers into the protocols section
Also makes the normal record header attribute names match their masks; previously they were their "understood" names.

The decision to move the messages into a 'decoded' subset of the protocols section is pending, but likely, as it would more reflect the 'raw' decoded messages.

The intent is to use the 'messages' subset of the protocols section for human readable/programmable works.