Refactors base type handling in FIT records from files.
I noticed that the base_type_field (number) is the combination of has_endianness and the base_type_number, and felt there was a better way to handle it.

Whether or not others like it, I like it much better now.

Was working on effectively useless unit tests, when some water decided to jump the gap between the sky and the ground, so not yet finished.

And "effectively useless" at this time, given that the base types are just an enumerated type.

However, there exists a chance that an additional method/function would be added to evaluate a field's bytes as read in a data message record.