Adds new pseudo 'FitMessageRecord' for associating definition and data records (#3)
* Adds new pseudo 'FitMessageRecord' for associating definition and data records

May have also started digging into a little more than the branch calls for...

Hi. I do that some times. It's a refactoring habit I can't _and_ won't stop doing. ^_^

* Refactors encoded (binary) record headers into the protocols section

Also makes the normal record header attribute names match their masks; previously they were their "understood" names.

The decision to move the messages into a 'decoded' subset of the protocols section is pending, but likely, as it would more reflect the 'raw' decoded messages.

The intent is to use the 'messages' subset of the protocols section for human readable/programmable works.

* Tweaks NRH and CTRH decode valations; adds NRH and CTRH unit tests

* Refactors base type handling in FIT records from files.

I noticed that the base_type_field (number) is the combination of has_endianness and the base_type_number, and felt there was a better way to handle it.

Whether or not others like it, I like it much better now.

Was working on effectively useless unit tests, when some water decided to jump the gap between the sky and the ground, so not yet finished.

And "effectively useless" at this time, given that the base types are just an enumerated type.

However, there exists a chance that an additional method/function would be added to evaluate a field's bytes as read in a data message record.

* Adds remaining unit tests for encoded base types

Also found a slight bug in that BaseType::name had two definitions; the latter was fixed to be the expected BaseType::invalid_value.

* [WIP] Fiels with evaluated values based on base_type

* Finishes adding evaluation functions for base types.

Begins basic unit tests for said evaluation functions, specifically for signed integers.

I noticed that the test file names aren't quite right for UInt* and UInt*Z classes (the file name is test_signed_uint*...).

* Fixes errant test file names

* Adds unit tests for unsigned integer base types

* Adds null-termination tests in preparation for String handling.

"Fixes" the usage of 'bytes' in the function parameters, now 'bites'. Type keywords, eh? ^_^

Cleans up some logic around testing invalid values on bytes.

Starts some floating point tests (the invalid values). Left a TODO for testing select/interesting values with links to Wikipedia for each (f/single, d/double).

* Moves combined def and data message records type

* Adds missed files intended for last commit

---------

Co-authored-by: Ms Pell Ted <misspellted@rouge-truth>